Partial clipboard attempt and bugged keylogger
Working on the clipboard functionalitity right now and am also trying to get the keylogger to work properly. The current problem is that it saves all the data from the start of the recording in every single line of the file instead of only logging the new data. E.g. if u typed 'hi' on Discord and then 'bye' on Skype, the first line will contain 'hi', while the second will contain 'hi bye', but it should say 'bye' only on the second line.