Fixed timestamps mostly
Only problem now is that I see no way of reading the text_file without any problems regarding the input text happening. If a user types [CLIPBOARD_DATA] anything after that will be used for future queries, meaning we can get problems. Need to find something which allows input from the program, but not from the user. Something only we can add that users can't add in any possible way. Only solution I can think of rn is adding some kind of key or word at the start of every line which allows the program to 'know' when the user input starts and which part is input by the program.